requirements.txt
Django>=3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
django-restql==0.14.0
djangorestframework==3.12.4
psycopg2-binary>=2.9